+++ src/app/common/services/storage/storage.service.ts
@@ -1,31 +1,34 @@
-import { Injectable } from '@angular/core';
+import { inject, Injectable } from '@angular/core';
+import { WINDOW } from 'src/app/core/platform/window.token';
 import { StorageKey } from './storage';
 
 @Injectable({
   providedIn: 'root',
 })
 export class StorageService implements Storage {
+  readonly window: Window | null = inject(WINDOW);
+
   getItem(key: StorageKey): string | null {
-    return localStorage.getItem(key);
+    return this.window && this.window.localStorage.getItem(key);
   }
 
   setItem(key: StorageKey, value: string): void {
-    return localStorage.setItem(key, value);
+    return this.window?.localStorage.setItem(key, value);
   }
 
   removeItem(key: StorageKey): void {
-    localStorage.removeItem(key);
+    this.window?.localStorage.removeItem(key);
   }
 
   clear(): void {
-    localStorage.clear();
+    this.window?.localStorage.clear();
   }
 
   key(index: number): string | null {
-    return localStorage.key(index);
+    return this.window && this.window.localStorage.key(index);
   }
 
   get length(): number {
-    return localStorage.length;
+    return this.window?.localStorage.length || 0;
   }
 }