+++ src/app/common/services/theme.service.ts
@@ -1,5 +1,5 @@
-import { Injectable, Signal, WritableSignal, computed, inject, signal } from '@angular/core';
-import { GoogleAnalyticsService } from 'src/app/common/services/google-analytics.service';
+import { Injectable, Signal, WritableSignal, afterNextRender, computed, inject, signal } from '@angular/core';
+import { AnalyticsService } from 'src/app/common/services/analytics/analytics.service';
 import { StorageKey } from 'src/app/common/services/storage/storage';
 import { StorageService } from 'src/app/common/services/storage/storage.service';
 
@@ -7,17 +7,19 @@ import { StorageService } from 'src/app/common/services/storage/storage.service'
   providedIn: 'root',
 })
 export class ThemeService {
-  private readonly googleAnalyticsService = inject(GoogleAnalyticsService);
+  private readonly analyticsService = inject(AnalyticsService);
   private readonly storageService = inject(StorageService);
 
   readonly #darkMode: WritableSignal<boolean> = signal(true);
   readonly darkMode: Signal<boolean> = computed(this.#darkMode);
 
   constructor() {
-    const darkModeSetting = this.storageService.getItem(StorageKey.DarkMode);
+    afterNextRender(() => {
+      const darkModeSetting = this.storageService.getItem(StorageKey.DarkMode);
 
-    if (!darkModeSetting) this.#darkMode.set(matchMedia('(prefers-color-scheme: dark)').matches);
-    else this.#darkMode.set(darkModeSetting === 'true');
+      if (darkModeSetting === null) this.#darkMode.set(matchMedia('(prefers-color-scheme: dark)').matches);
+      else this.#darkMode.set(darkModeSetting === 'true');
+    });
   }
 
   toggleDarkMode() {
@@ -25,7 +27,7 @@ export class ThemeService {
 
     this.storageService.setItem(StorageKey.DarkMode, JSON.stringify(this.darkMode()));
 
-    this.googleAnalyticsService.trackEvent('toggle_dark_mode', 'theming', 'dark_mode', this.darkMode());
+    this.analyticsService.trackEvent('toggle_dark_mode', 'theming', 'dark_mode', this.darkMode());
 
     document.documentElement.classList.toggle('dark', this.darkMode());
   }