+++ src/app/core/error-handling/error-handler.ts
@@ -1,17 +1,17 @@
 import { HttpErrorResponse } from '@angular/common/http';
 import { ErrorHandler, Injectable, inject } from '@angular/core';
-import { GoogleAnalyticsService } from 'src/app/common/services/google-analytics.service';
+import { AnalyticsService } from 'src/app/common/services/analytics/analytics.service';
 
 @Injectable()
 export class CustomErrorHandler implements ErrorHandler {
-  private readonly googleAnalyticsService = inject(GoogleAnalyticsService);
+  private readonly analyticsService = inject(AnalyticsService);
 
   handleError(error: Error) {
     if (error instanceof HttpErrorResponse) {
       // eslint-disable-next-line no-console
       console.error(error);
     } else {
-      this.googleAnalyticsService.trackException(error.message, true, error);
+      this.analyticsService.trackException(error.message, true, error);
     }
   }
 }