+++ src/app/features/trackers/xp-tracker/player-detail/player-detail-widget/player-detail-widget.component.ts
@@ -3,7 +3,7 @@ import { Component, InputSignal, inject, input } from '@angular/core';
 import { Player, PlayerStatus, PlayerType } from '@osrs-tracker/models';
 import { IconDirective } from 'src/app/common/directives/icon/icon.directive';
 import { CapitalizePipe } from 'src/app/common/pipes/capitalize.pipe';
-import { GoogleAnalyticsService } from 'src/app/common/services/google-analytics.service';
+import { AnalyticsService } from 'src/app/common/services/analytics/analytics.service';
 import { XpTrackerStorageService } from '../../xp-tracker-storage.service';
 
 @Component({
@@ -12,7 +12,7 @@ import { XpTrackerStorageService } from '../../xp-tracker-storage.service';
   imports: [CapitalizePipe, DatePipe, IconDirective],
 })
 export class PlayerDetailWidgetComponent {
-  private readonly googlAnalyticsService = inject(GoogleAnalyticsService);
+  private readonly analyticsService = inject(AnalyticsService);
   private readonly xpTrackerStorageService = inject(XpTrackerStorageService);
 
   readonly PlayerType: typeof PlayerType = PlayerType;
@@ -27,7 +27,7 @@ export class PlayerDetailWidgetComponent {
   toggleFavorite(): void {
     this.xpTrackerStorageService.toggleFavoritePlayer(this.playerDetail().username);
 
-    this.googlAnalyticsService.trackEvent(
+    this.analyticsService.trackEvent(
       'toggle_favorite_player',
       'xp_tracker',
       this.playerDetail().username,